Remove mark objects in favor of a stack of arrays
The recursive descent parser implemented here follows a certain pattern
that backtracks by marking the stream position using Mark objects. These
were allocated on the heap and thrown away.

Given the high frequency with which the stream position is marked for
possible backtracking, this generated a large number of throwaway Mark
objects.

We replace this with a single stack of reusable int[] which hold the same 4
fields: stream offset, line / char offset, and flags.